Controller da finire

+++ src/main/java/com/inbook/service/BookService.java
@@ -11,7 +11,7 @@ public class BookService {
 
     public BookService(BookRepository repo) { this.repo = repo; }
 
-    public BookRepository addBook(String isbn, String autore, String titolo, int volume, String casaEditrice, int prezzo, boolean daAcquistare, boolean consigliato){
+    public BookRepository addBook(String isbn, String autore, String titolo, int volume, String casaEditrice, double prezzo, boolean daAcquistare, boolean consigliato){
 
         Book book = new Book();
